requirements.txt
streamlit>=1.35.0
numpy>=1.26.4
Pillow>=10.3.0
torch>=2.4.0
torchvision>=0.19.0
tensorflow>=2.17.0
protobuf>=3.20.0
matplotlib>=3.9.2
requests>=2.32.0
